+++ LyranetworkPayzen/src/Service/OrderService.php
@@ -14,6 +14,14 @@ namespace Lyranetwork\Payzen\Service;
 
 use Lyranetwork\Payzen\Repository\OrderRepositoryInterface;
 
+use Sylius\Bundle\ShopBundle\EmailManager\OrderEmailManagerInterface;
+use App\Entity\Payment\Payment;
+use Symfony\Component\HttpClient\Exception\TransportException;
+
+use Webmozart\Assert\Assert;
+
+use Psr\Log\LoggerInterface;
+
 class OrderService
 {
     /**
@@ -21,11 +29,25 @@ class OrderService
      */
     private $orderRepository;
 
+    /**
+     * @var OrderEmailManagerInterface
+     */
+    private $orderEmailManager;
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
     public function __construct(
-        OrderRepositoryInterface $orderRepository
+        OrderRepositoryInterface $orderRepository,
+        OrderEmailManagerInterface $orderEmailManager,
+        LoggerInterface $logger
     )
     {
         $this->orderRepository = $orderRepository;
+        $this->orderEmailManager = $orderEmailManager;
+        $this->logger = $logger;
     }
 
     public function get(string $orderId)
@@ -36,5 +58,24 @@ class OrderService
     public function getByNumber(string $orderNumber)
     {
         return $this->orderRepository->findOneByNumber($orderNumber);
+    }
+
+    public function sendConfirmationEmail(Payment $payment): void
+    {
+        try {
+            Assert::isInstanceOf($payment, Payment::class);
+
+            $gatewayName = constant('Lyranetwork\Payzen\Payum\SyliusPaymentGatewayFactory::FACTORY_NAME');
+            $factoryName = $payment->getMethod()->getGatewayConfig()->getFactoryName();
+            if ($gatewayName === $factoryName) {
+                $order = $payment->getOrder();
+
+                $this->logger->info("Sending confirmation email for order: " . $order->getNumber());
+
+                $this->orderEmailManager->sendConfirmationEmail($order);
+            }
+        } catch (TransportException $e) {
+            $this->logger->error($e->getMessage());
+        }
     }
 }