+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.2, 2024-08-06:
+- Bug fix: Fix error related to checkout state update at the end of the order.
+- Bug fix: Fix error related to sending confirmation email.
+
 # 1.1.1, 2024-07-26:
 - Bug fix: Fix error related to form validation.
 - Improve translations management.